+++ src/Entities/Update.php
@@ -38,27 +38,35 @@ use Longman\TelegramBot\Entities\Payments\ShippingQuery;
  * @method ChatJoinRequest             getChatJoinRequest()      Optional. A request to join the chat has been sent. The bot must have the can_invite_users administrator right in the chat to receive these updates.
  * @method ChatBoostUpdated            getChatBoost()            Optional. A chat boost was added or changed. The bot must be an administrator in the chat to receive these updates.
  * @method ChatBoostRemoved            getRemovedChatBoost()     Optional. A boost was removed from a chat. The bot must be an administrator in the chat to receive these updates.
+ * @method BusinessConnection          getBusinessConnection()     Optional. The bot was connected to or disconnected from a business account, or a user edited an existing connection with the bot
+ * @method Message                     getBusinessMessage()        Optional. New message from a connected business account
+ * @method Message                     getEditedBusinessMessage()  Optional. New version of a message from a connected business account
+ * @method BusinessMessagesDeleted     getDeletedBusinessMessages() Optional. Messages were deleted from a connected business account
  */
 class Update extends Entity
 {
-    public const TYPE_MESSAGE                = 'message';
-    public const TYPE_EDITED_MESSAGE         = 'edited_message';
-    public const TYPE_CHANNEL_POST           = 'channel_post';
-    public const TYPE_EDITED_CHANNEL_POST    = 'edited_channel_post';
-    public const TYPE_MESSAGE_REACTION       = 'message_reaction';
-    public const TYPE_MESSAGE_REACTION_COUNT = 'message_reaction_count';
-    public const TYPE_INLINE_QUERY           = 'inline_query';
-    public const TYPE_CHOSEN_INLINE_RESULT   = 'chosen_inline_result';
-    public const TYPE_CALLBACK_QUERY         = 'callback_query';
-    public const TYPE_SHIPPING_QUERY         = 'shipping_query';
-    public const TYPE_PRE_CHECKOUT_QUERY     = 'pre_checkout_query';
-    public const TYPE_POLL                   = 'poll';
-    public const TYPE_POLL_ANSWER            = 'poll_answer';
-    public const TYPE_MY_CHAT_MEMBER         = 'my_chat_member';
-    public const TYPE_CHAT_MEMBER            = 'chat_member';
-    public const TYPE_CHAT_JOIN_REQUEST      = 'chat_join_request';
-    public const TYPE_CHAT_BOOST             = 'chat_boost';
-    public const TYPE_REMOVED_CHAT_BOOST     = 'removed_chat_boost';
+    public const TYPE_MESSAGE                  = 'message';
+    public const TYPE_EDITED_MESSAGE           = 'edited_message';
+    public const TYPE_CHANNEL_POST             = 'channel_post';
+    public const TYPE_EDITED_CHANNEL_POST      = 'edited_channel_post';
+    public const TYPE_MESSAGE_REACTION         = 'message_reaction';
+    public const TYPE_MESSAGE_REACTION_COUNT   = 'message_reaction_count';
+    public const TYPE_INLINE_QUERY             = 'inline_query';
+    public const TYPE_CHOSEN_INLINE_RESULT     = 'chosen_inline_result';
+    public const TYPE_CALLBACK_QUERY           = 'callback_query';
+    public const TYPE_SHIPPING_QUERY           = 'shipping_query';
+    public const TYPE_PRE_CHECKOUT_QUERY       = 'pre_checkout_query';
+    public const TYPE_POLL                     = 'poll';
+    public const TYPE_POLL_ANSWER              = 'poll_answer';
+    public const TYPE_MY_CHAT_MEMBER           = 'my_chat_member';
+    public const TYPE_CHAT_MEMBER              = 'chat_member';
+    public const TYPE_CHAT_JOIN_REQUEST        = 'chat_join_request';
+    public const TYPE_CHAT_BOOST               = 'chat_boost';
+    public const TYPE_REMOVED_CHAT_BOOST       = 'removed_chat_boost';
+    public const TYPE_BUSINESS_CONNECTION      = 'business_connection';
+    public const TYPE_BUSINESS_MESSAGE         = 'business_message';
+    public const TYPE_EDITED_BUSINESS_MESSAGE  = 'edited_business_message';
+    public const TYPE_DELETED_BUSINESS_MESSAGES = 'deleted_business_messages';
 
     /**
      * {@inheritdoc}
@@ -66,24 +74,28 @@ class Update extends Entity
     protected function subEntities(): array
     {
         return [
-            self::TYPE_MESSAGE                => Message::class,
-            self::TYPE_EDITED_MESSAGE         => EditedMessage::class,
-            self::TYPE_CHANNEL_POST           => ChannelPost::class,
-            self::TYPE_EDITED_CHANNEL_POST    => EditedChannelPost::class,
-            self::TYPE_MESSAGE_REACTION       => MessageReactionUpdated::class,
-            self::TYPE_MESSAGE_REACTION_COUNT => MessageReactionCountUpdated::class,
-            self::TYPE_INLINE_QUERY           => InlineQuery::class,
-            self::TYPE_CHOSEN_INLINE_RESULT   => ChosenInlineResult::class,
-            self::TYPE_CALLBACK_QUERY         => CallbackQuery::class,
-            self::TYPE_SHIPPING_QUERY         => ShippingQuery::class,
-            self::TYPE_PRE_CHECKOUT_QUERY     => PreCheckoutQuery::class,
-            self::TYPE_POLL                   => Poll::class,
-            self::TYPE_POLL_ANSWER            => PollAnswer::class,
-            self::TYPE_MY_CHAT_MEMBER         => ChatMemberUpdated::class,
-            self::TYPE_CHAT_MEMBER            => ChatMemberUpdated::class,
-            self::TYPE_CHAT_JOIN_REQUEST      => ChatJoinRequest::class,
-            self::TYPE_CHAT_BOOST             => ChatBoostUpdated::class,
-            self::TYPE_REMOVED_CHAT_BOOST     => ChatBoostRemoved::class,
+            self::TYPE_MESSAGE                  => Message::class,
+            self::TYPE_EDITED_MESSAGE           => EditedMessage::class,
+            self::TYPE_CHANNEL_POST             => ChannelPost::class,
+            self::TYPE_EDITED_CHANNEL_POST      => EditedChannelPost::class,
+            self::TYPE_MESSAGE_REACTION         => MessageReactionUpdated::class,
+            self::TYPE_MESSAGE_REACTION_COUNT   => MessageReactionCountUpdated::class,
+            self::TYPE_INLINE_QUERY             => InlineQuery::class,
+            self::TYPE_CHOSEN_INLINE_RESULT     => ChosenInlineResult::class,
+            self::TYPE_CALLBACK_QUERY           => CallbackQuery::class,
+            self::TYPE_SHIPPING_QUERY           => ShippingQuery::class,
+            self::TYPE_PRE_CHECKOUT_QUERY       => PreCheckoutQuery::class,
+            self::TYPE_POLL                     => Poll::class,
+            self::TYPE_POLL_ANSWER              => PollAnswer::class,
+            self::TYPE_MY_CHAT_MEMBER           => ChatMemberUpdated::class,
+            self::TYPE_CHAT_MEMBER              => ChatMemberUpdated::class,
+            self::TYPE_CHAT_JOIN_REQUEST        => ChatJoinRequest::class,
+            self::TYPE_CHAT_BOOST               => ChatBoostUpdated::class,
+            self::TYPE_REMOVED_CHAT_BOOST       => ChatBoostRemoved::class,
+            self::TYPE_BUSINESS_CONNECTION      => BusinessConnection::class,
+            self::TYPE_BUSINESS_MESSAGE         => Message::class,
+            self::TYPE_EDITED_BUSINESS_MESSAGE  => Message::class,
+            self::TYPE_DELETED_BUSINESS_MESSAGES => BusinessMessagesDeleted::class,
         ];
     }